+++ src/Container.ts
@@ -20,6 +20,24 @@ export class Container implements ContainerInterface {
   protected _fakes: BindingMapInterface = new BindingMap()
 
   /**
+   * Retrieve the value of the bindings property.
+   *
+   * @return {BindingMapInterface}
+   */
+  public get bindings (): BindingMapInterface {
+    return this._bindings
+  }
+
+  /**
+   * Retrieve the value of the fakes property.
+   *
+   * @return {BindingMapInterface}
+   */
+  public get fakes (): BindingMapInterface {
+    return this._fakes
+  }
+
+  /**
    * Bind a single instance or value into the container under the provided key.
    *
    * @param  {any} key
@@ -164,23 +182,5 @@ export class Container implements ContainerInterface {
     }
 
     return this
-  }
-
-  /**
-   * Retrieve the value of the bindings property.
-   *
-   * @return {BindingMapInterface}
-   */
-  public getBindings (): BindingMapInterface {
-    return this._bindings
-  }
-
-  /**
-   * Retrieve the value of the fakes property.
-   *
-   * @return {BindingMapInterface}
-   */
-  public getFakes (): BindingMapInterface {
-    return this._fakes
   }
 }