+++ src/bindings/Binding.ts
@@ -3,9 +3,9 @@ import { ContainerInterface } from '../interfaces/ContainerInterface'
 
 export abstract class Binding implements BindingInterface {
   /**
-   * @var {any} _value
+   * @var {any} value
    */
-  protected _value: any = null
+  public value: any = null
 
   /**
    * Create a new instance of Binding.
@@ -13,27 +13,7 @@ export abstract class Binding implements BindingInterface {
    * @param {any = null}  value
    */
   public constructor (value: any = null) {
-    this._value = value
-  }
-
-  /**
-   * Retrieve the value of the value property.
-   *
-   * @return {any}
-   */
-  public getValue (): any {
-    return this._value
-  }
-
-  /**
-   * Set the value of the value property.
-   *
-   * @param  {any}  value
-   * @return {this}
-   */
-  public setValue (value: any): this {
-    this._value = value
-    return this
+    this.value = value
   }
 
   /**