+++ src/bindings/ResolverBinding.ts
@@ -4,14 +4,23 @@ import { ResolverBindingInterface } from '../interfaces/bindings/ResolverBinding
 
 export abstract class ResolverBinding extends Binding implements ResolverBindingInterface {
   /**
-   * @var {boolean} _hasResolved
+   * @var {boolean} hasResolved
    */
-  protected _hasResolved: boolean = false
+  public hasResolved: boolean = false
 
   /**
    * @var {Resolver|null} _resolver
    */
   protected _resolver: Resolver
+
+  /**
+   * Retrieve the value of the resolver property.
+   *
+   * @return {Resolver}
+   */
+  get resolver (): Resolver {
+    return this._resolver
+  }
 
   /**
    * Create a new instance of ResolverBinding.
@@ -22,34 +31,5 @@ export abstract class ResolverBinding extends Binding implements ResolverBinding
     super()
 
     this._resolver = resolver
-  }
-
-  /**
-   * Retrieve the value of the resolver property.
-   *
-   * @return {Resolver}
-   */
-  public getResolver (): Resolver {
-    return this._resolver
-  }
-
-  /**
-   * Determine if the binding has been resolved at least once.
-   *
-   * @return {boolean}
-   */
-  public hasResolved (): boolean {
-    return this._hasResolved
-  }
-
-  /**
-   * Set the value of the hasResolved property.
-   *
-   * @param  {boolean} hasResolved
-   * @return {this}
-   */
-  public setHasResolved (hasResolved: boolean): this {
-    this._hasResolved = hasResolved
-    return this
   }
 }